+++ src/azure-cli/azure/cli/command_modules/network/commands.py
@@ -3,7 +3,7 @@
 # Licensed under the MIT License. See License.txt in the project root for license information.
 # --------------------------------------------------------------------------------------------
 
-# pylint: disable=line-too-long
+# pylint: disable=line-too-long,too-many-lines
 
 from azure.cli.core.commands import DeploymentOutputLongRunningOperation
 from azure.cli.core.commands.arm import (
@@ -26,7 +26,8 @@ from azure.cli.command_modules.network._client_factory import (
     cf_service_endpoint_policy_definitions, cf_dns_references, cf_private_endpoints, cf_network_profiles,
     cf_express_route_circuit_connections, cf_express_route_gateways, cf_express_route_connections,
     cf_express_route_ports, cf_express_route_port_locations, cf_express_route_links, cf_app_gateway_waf_policy,
-    cf_service_tags, cf_private_link_services, cf_private_endpoint_types, cf_peer_express_route_circuit_connections)
+    cf_service_tags, cf_private_link_services, cf_private_endpoint_types, cf_peer_express_route_circuit_connections,
+    cf_virtual_router, cf_virtual_router_peering)
 from azure.cli.command_modules.network._util import (
     list_network_resource_property, get_network_resource_property_entry, delete_network_resource_property_entry)
 from azure.cli.command_modules.network._format import (
@@ -313,6 +314,18 @@ def load_command_table(self, _):
         min_api='2018-07-01'
     )
 
+    network_vrouter_sdk = CliCommandType(
+        operations_tmpl='azure.mgmt.network.operations#VirtualRoutersOperations.{}',
+        client_factory=cf_virtual_router,
+        min_api='2019-08-01'
+    )
+
+    network_vrouter_peering_sdk = CliCommandType(
+        operations_tmpl='azure.mgmt.network.operations#VirtualRouterPeeringsOperations.{}',
+        client_factory=cf_virtual_router_peering,
+        min_api='2019-08-01'
+    )
+
     network_custom = CliCommandType(operations_tmpl='azure.cli.command_modules.network.custom#{}')
 
     # endregion
@@ -545,7 +558,8 @@ def load_command_table(self, _):
         g.wait_command('wait')
 
     with self.command_group('network express-route auth', network_erca_sdk) as g:
-        g.command('create', 'create_or_update', validator=process_auth_create_namespace)
+        g.custom_command('create', 'create_express_route_auth', min_api='2019-09-01')
+        g.command('create', 'create_or_update', max_api='2019-08-01', validator=process_auth_create_namespace)
         g.command('delete', 'delete')
         g.show_command('show', 'get')
         g.command('list', 'list')
@@ -986,4 +1000,20 @@ def load_command_table(self, _):
         g.custom_command('add', 'add_vpn_conn_ipsec_policy', supports_no_wait=True, doc_string_source='IpsecPolicy')
         g.custom_command('list', 'list_vpn_conn_ipsec_policies')
         g.custom_command('clear', 'clear_vpn_conn_ipsec_policies', supports_no_wait=True)
+    # endregion
+
+    # region VirtualRouter
+    with self.command_group('network vrouter', network_vrouter_sdk) as g:
+        g.custom_command('create', 'create_virtual_router')
+        g.generic_update_command('update', custom_func_name='update_virtual_router')
+        g.command('delete', 'delete')
+        g.show_command('show', 'get')
+        g.custom_command('list', 'list_virtual_router')
+
+    with self.command_group('network vrouter peering', network_vrouter_peering_sdk) as g:
+        g.custom_command('create', 'create_virtual_router_peering')
+        g.generic_update_command('update', custom_func_name='update_virtual_router_peering')
+        g.command('delete', 'delete')
+        g.show_command('show', 'get')
+        g.command('list', 'list')
     # endregion